requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backoff==2.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
Deprecated==1.2.18
distro==1.9.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
langfuse==2.60.5
limits==5.1.0
openai==1.78.1
packaging==24.2
pluggy==1.6.0
pydantic==2.11.4
pydantic_core==2.33.2
pytest==8.3.5
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
slowapi==0.1.9
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
